move page visibility back into track
