+++ src/room/track/RemoteVideoTrack.ts
@@ -14,7 +14,6 @@ import { attachToElement, detachTrack, Track } from './Track';
 import type { AdaptiveStreamSettings } from './types';
 
 const REACTION_DELAY = 100;
-const BACKGROUND_REACTION_DELAY = 5000;
 
 export default class RemoteVideoTrack extends RemoteTrack {
   private prevStats?: VideoReceiverStats;
@@ -28,10 +27,6 @@ export default class RemoteVideoTrack extends RemoteTrack {
   private lastDimensions?: Track.Dimensions;
 
   private hasUsedAttach: boolean = false;
-
-  protected isInBackground: boolean = false;
-
-  private backgroundTimeout: ReturnType<typeof setTimeout> | undefined;
 
   constructor(
     mediaTrack: MediaStreamTrack,
@@ -73,9 +68,6 @@ export default class RemoteVideoTrack extends RemoteTrack {
   attach(): HTMLMediaElement;
   attach(element: HTMLMediaElement): HTMLMediaElement;
   attach(element?: HTMLMediaElement): HTMLMediaElement {
-    if (this.attachedElements.length === 0) {
-      this.addPageVisibilityListener();
-    }
     if (!element) {
       element = super.attach();
     } else {
@@ -144,24 +136,18 @@ export default class RemoteVideoTrack extends RemoteTrack {
   detach(): HTMLMediaElement[];
   detach(element: HTMLMediaElement): HTMLMediaElement;
   detach(element?: HTMLMediaElement): HTMLMediaElement | HTMLMediaElement[] {
-    try {
-      let detachedElements: HTMLMediaElement[] = [];
-      if (element) {
-        this.stopObservingElement(element);
-        return super.detach(element);
-      }
-      detachedElements = super.detach();
-
-      for (const e of detachedElements) {
-        this.stopObservingElement(e);
-      }
-
-      return detachedElements;
-    } finally {
-      if (this.attachedElements.length === 0) {
-        this.removePageVisibilityListener();
-      }
+    let detachedElements: HTMLMediaElement[] = [];
+    if (element) {
+      this.stopObservingElement(element);
+      return super.detach(element);
     }
+    detachedElements = super.detach();
+
+    for (const e of detachedElements) {
+      this.stopObservingElement(e);
+    }
+
+    return detachedElements;
   }
 
   /** @internal */
@@ -223,7 +209,7 @@ export default class RemoteVideoTrack extends RemoteTrack {
   }
 
   protected async handleAppVisibilityChanged() {
-    this.isInBackground = document.visibilityState === 'hidden';
+    await super.handleAppVisibilityChanged();
     if (!this.isAdaptiveStream) return;
     this.updateVisibility();
   }
@@ -287,37 +273,6 @@ export default class RemoteVideoTrack extends RemoteTrack {
 
     this.emit(TrackEvent.VideoDimensionsChanged, this.lastDimensions, this);
   }
-
-  protected addPageVisibilityListener() {
-    if (isWeb()) {
-      this.isInBackground = document.visibilityState === 'hidden';
-      document.addEventListener('visibilitychange', this.pageVisibilityChangedListener);
-    } else {
-      this.isInBackground = false;
-    }
-  }
-
-  protected removePageVisibilityListener() {
-    if (isWeb()) {
-      document.removeEventListener('visibilitychange', this.pageVisibilityChangedListener);
-    }
-  }
-
-  protected pageVisibilityChangedListener = () => {
-    if (this.backgroundTimeout) {
-      clearTimeout(this.backgroundTimeout);
-    }
-    // delay app visibility update if it goes to hidden
-    // update immediately if it comes back to focus
-    if (document.visibilityState === 'hidden') {
-      this.backgroundTimeout = setTimeout(
-        () => this.handleAppVisibilityChanged(),
-        BACKGROUND_REACTION_DELAY,
-      );
-    } else {
-      this.handleAppVisibilityChanged();
-    }
-  };
 }
 
 export interface ElementInfo {
